templates: Move *MessageHelper from common to stub
issues/55 has justifiable questions the necessity of having *MessageHelper
classes on the common. So as advices this commit moves and namespaces it to
*stub.h; this also adds namespaces for *TypeWrap classes.
