+++ README.md
@@ -112,9 +112,9 @@ C++ class. In our example, the following will suffice:
 class BarImpl : public org::foo::Bar {
 public:
     // Called wben org.foo.bar.Baz() is invoked
-    void Baz (BarMessageHelper msg) override {
+    void Baz (MethodInvocation invocation) override {
         // Return void
-        msg.ret();
+        invocation.ret();
     }
 };
