+++ tests/data/simple/input_common.h
@@ -1,9 +1,13 @@
 #pragma once
 #include <iostream>
-#include <map>
-#include <tuple>
+#include <vector>
 #include "glibmm.h"
 #include "giomm.h"
+
+namespace org {
+namespace gdbus {
+namespace codegen {
+namespace glibmm {
 
 class TestTypeWrap {
 public:
@@ -51,38 +55,10 @@ public:
     }
 };
 
-class TestMessageHelper {
-public:
-    TestMessageHelper (const Glib::RefPtr<Gio::DBus::MethodInvocation> msg):
-        m_message(msg) {}
-
-    const Glib::RefPtr<Gio::DBus::MethodInvocation> getMessage() {
-        return m_message;
-    }
-
-    void ret(Glib::Error error) {
-        m_message->return_error(error);
-    }
-
-    void returnError(const Glib::ustring &domain, int code, const Glib::ustring &message) {
-        m_message->return_error(domain, code, message);
-    }
-
-    void ret(const Glib::ustring & p0, const std::map<Glib::ustring,Glib::VariantBase> & p1) {
-        std::vector<Glib::VariantBase> vlist;
-        Glib::Variant<Glib::ustring> var0 =
-            Glib::Variant<Glib::ustring>::create(p0);
-        vlist.push_back(var0);
-        Glib::Variant<std::map<Glib::ustring,Glib::VariantBase>> var1 =
-            Glib::Variant<std::map<Glib::ustring,Glib::VariantBase>>::create(p1);
-        vlist.push_back(var1);
-
-        m_message->return_value(Glib::Variant<Glib::VariantBase>::create_tuple(vlist));
-    }
-
-private:
-    Glib::RefPtr<Gio::DBus::MethodInvocation> m_message;
-};
+} // glibmm
+} // codegen
+} // gdbus
+} // org
 
 namespace org {
 namespace gdbus {
